Split off layouts into separate files; added modern_nodesc layout
