+++ src/scss/_layout-modern.scss
@@ -1,4 +1,5 @@
-.ucf-events-modern {
+.ucf-events-modern,
+.ucf-events-modern-nodesc {
 
   .ucf-event-row-margin {
     margin-bottom: 1.7rem;
