+++ ucf-events.php
@@ -19,6 +19,9 @@ require_once 'includes/ucf-events-common.php';
 require_once 'includes/ucf-events-shortcode.php';
 require_once 'includes/ucf-events-widget.php';
 
+require_once 'layouts/ucf-events-classic.php';
+require_once 'layouts/ucf-events-modern.php';
+
 
 /**
  * Activation/deactivation hooks
